chore: drop pre-1.21.8 support
