+++ gradle.properties
@@ -1,6 +1,6 @@
 version=1.0.0
 group=top.tobyprime
-supported_mc_versions=1.21.1,1.21.2,1.21.5,1.21.6,1.21.9,1.21.11
+supported_mc_versions=1.21.8,1.21.9,1.21.11,26.1
 supported_native_platforms=windows-x86_64,linux-x86_64,linux-arm64,macosx-x86_64,macosx-arm64
 
 gson_version=2.13.2
