--- mcedia-cn-platforms/versions/1.21.1/gradle.properties
@@ -1,4 +1,0 @@
-minecraft_version=1.21.1
-loader_version=0.16.14
-fabric_api_version=0.116.12+1.21.1
-java_version=21
