--- mcedia-cn-platforms/versions/1.21.2/gradle.properties
@@ -1,5 +1,0 @@
-minecraft_version=1.21.2
-minecraft_version_range=>=1.21.2 <=1.21.4
-loader_version=0.16.14
-fabric_api_version=0.114.1+1.21.3
-java_version=21
