--- mcedia-cn-platforms/versions/1.21.6/gradle.properties
@@ -1,5 +1,0 @@
-minecraft_version=1.21.6
-minecraft_version_range=>=1.21.6 <=1.21.8
-loader_version=0.16.14
-fabric_api_version=0.128.2+1.21.6
-java_version=21
