+++ mcedia-cn-platforms/versions/1.21.8/gradle.properties
@@ -1,0 +1,5 @@
+minecraft_version=1.21.8
+minecraft_version_range=1.21.8
+loader_version=0.17.3
+fabric_api_version=0.136.1+1.21.8
+java_version=21
