--- mcedia-core/versions/1.21.1/src/client/java/top/tobyprime/mcedia_core/client/debug/DebugRendererAudioHitbox.java
@@ -1,14 +1,0 @@
-package top.tobyprime.mcedia_core.client.debug;
-
-import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.renderer.SubmitNodeCollector;
-import net.minecraft.client.renderer.culling.Frustum;
-import net.minecraft.client.renderer.state.CameraRenderState;
-
-public final class DebugRendererAudioHitbox {
-    private DebugRendererAudioHitbox() {
-    }
-
-    public static void submit(PoseStack poseStack, SubmitNodeCollector submitNodeCollector, CameraRenderState cameraRenderState, Frustum frustum) {
-    }
-}
