--- mcedia-core/versions/1.21.1/src/client/java/top/tobyprime/mcedia_core/client/mixin/MixinDebugRenderer.java
@@ -1,8 +1,0 @@
-package top.tobyprime.mcedia_core.client.mixin;
-
-import net.minecraft.client.renderer.debug.DebugRenderer;
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(DebugRenderer.class)
-public class MixinDebugRenderer {
-}
