--- mcedia-core/versions/1.21.1/src/client/java/top/tobyprime/mcedia_core/client/mixin/MixinLevelRenderer.java
@@ -1,8 +1,0 @@
-package top.tobyprime.mcedia_core.client.mixin;
-
-import net.minecraft.client.renderer.LevelRenderer;
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(LevelRenderer.class)
-public class MixinLevelRenderer {
-}
