--- mcedia-core/versions/1.21.1/src/client/java/top/tobyprime/mcedia_core/client/renderer/McediaRenderer.java
@@ -1,118 +1,0 @@
-package top.tobyprime.mcedia_core.client.renderer;
-
-import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.LevelRenderer;
-import net.minecraft.client.renderer.culling.Frustum;
-import net.minecraft.core.BlockPos;
-import org.jspecify.annotations.Nullable;
-import top.tobyprime.mcedia_core.client.player.MediaPlayerPeripheral;
-import top.tobyprime.mcedia_core.client.player.PlayerHost;
-import top.tobyprime.mcedia_core.client.player.ScreenPeripheral;
-
-import java.util.Collections;
-import java.util.LinkedHashSet;
-import java.util.Set;
-
-public final class McediaRenderer {
-    private static final McediaRenderer INSTANCE = new McediaRenderer();
-
-    private final Object lock = new Object();
-    private final Set<ScreenPeripheral> screens = new LinkedHashSet<>();
-    private final Set<MediaPlayerPeripheral> peripherals = new LinkedHashSet<>();
-    private volatile @Nullable Frustum currentFrustum;
-
-    private McediaRenderer() {
-    }
-
-    public static McediaRenderer get() {
-        return INSTANCE;
-    }
-
-    public void registerPeripheral(PlayerHost host, MediaPlayerPeripheral peripheral) {
-        host.addPeripheral(peripheral);
-        if (!host.getPeripherals().contains(peripheral)) {
-            return;
-        }
-        synchronized (lock) {
-            peripherals.add(peripheral);
-            if (peripheral instanceof ScreenPeripheral screen) {
-                screens.add(screen);
-            }
-        }
-    }
-
-    public void unregisterPeripheral(MediaPlayerPeripheral peripheral) {
-        synchronized (lock) {
-            peripherals.remove(peripheral);
-            if (peripheral instanceof ScreenPeripheral screen) {
-                screens.remove(screen);
-            }
-        }
-    }
-
-    public Set<MediaPlayerPeripheral> getPeripherals() {
-        synchronized (lock) {
-            return Collections.unmodifiableSet(new LinkedHashSet<>(peripherals));
-        }
-    }
-
-    public Set<MediaPlayerPeripheral> getPeripherals(PlayerHost host) {
-        return host.getPeripherals();
-    }
-
-    public boolean hasActivePeripheral(PlayerHost host, top.tobyprime.mcedia_core.client.player.PeripheralType type) {
-        for (var peripheral : host.getPeripherals()) {
-            if (peripheral.getPeripheralType() == type && peripheral.isActive()) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public void cleanup() {
-        synchronized (lock) {
-            peripherals.removeIf(peripheral -> !peripheral.isAlive());
-            screens.removeIf(screen -> !screen.isAlive());
-        }
-    }
-
-    public void setCurrentFrustum(@Nullable Frustum frustum) {
-        this.currentFrustum = frustum;
-    }
-
-    public @Nullable Frustum getCurrentFrustum() {
-        return currentFrustum;
-    }
-
-    public void setCameraPosition(double x, double y, double z) {
-    }
-
-    public void submitScreens(PoseStack poseStack) {
-        var level = Minecraft.getInstance().level;
-        if (level == null) {
-            return;
-        }
-
-        var frustum = getCurrentFrustum();
-        for (var screen : snapshotScreens()) {
-            if (!screen.isAlive()) {
-                continue;
-            }
-            if (frustum != null && !screen.isVisible(frustum)) {
-                continue;
-            }
-
-            var state = PlayerScreenEntityRenderer.createRenderState();
-            int light = LevelRenderer.getLightColor(level, BlockPos.containing(screen.getPosition()));
-            PlayerScreenEntityRenderer.extractRenderState(screen, state, light);
-            PlayerScreenEntityRenderer.submit(state);
-        }
-    }
-
-    private Set<ScreenPeripheral> snapshotScreens() {
-        synchronized (lock) {
-            return new LinkedHashSet<>(screens);
-        }
-    }
-}
