--- mcedia-core/versions/1.21.1/src/client/resources/mcedia_core.client.mixins.json
@@ -1,25 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "top.tobyprime.mcedia_core.client.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "client": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "overwrites": {
-    "requireAnnotations": true
-  },
-  "mixins": [
-    "MixinGameRenderer",
-    "MixinLevelRenderer",
-    "MixinDebugRenderer",
-    "MixinSoundEngine",
-    "MixinSoundOptionsScreen",
-    "MixinDebugScreenEntries",
-    "accessor.SoundManagerAccessor",
-    "accessor.SoundEngineAccessor",
-    "accessor.OptionsSubScreenAccessor"
-  ]
-}
