--- mcedia-core/versions/1.21.5/gradle.properties
@@ -1,4 +1,0 @@
-minecraft_version=1.21.5
-loader_version=0.16.14
-fabric_api_version=0.128.1+1.21.5
-java_version=21
