--- mcedia-core/versions/1.21.5/src/client/java/top/tobyprime/mcedia_core/client/debug/DebugEntryMMDecoderMetrics.java
@@ -1,5 +1,0 @@
-package top.tobyprime.mcedia_core.client.debug;
-
-public class DebugEntryMMDecoderMetrics {
-    public static final String ENTRY_ID = "mcedia_mm_decoder_metrics";
-}
