--- mcedia-core/versions/1.21.5/src/client/java/top/tobyprime/mcedia_core/client/mixin/MixinDebugScreenEntries.java
@@ -1,7 +1,0 @@
-package top.tobyprime.mcedia_core.client.mixin;
-
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(net.minecraft.client.Minecraft.class)
-public class MixinDebugScreenEntries {
-}
