+++ mcedia-core/versions/1.21.8/src/client/java/top/tobyprime/mcedia_core/client/debug/DebugRendererAudioHitbox.java
@@ -1,0 +1,9 @@
+package top.tobyprime.mcedia_core.client.debug;
+
+public final class DebugRendererAudioHitbox {
+    private DebugRendererAudioHitbox() {
+    }
+
+    public static void init() {
+    }
+}
