+++ mcedia-core/versions/1.21.8/src/client/java/top/tobyprime/mcedia_core/client/mixin/MixinDebugScreenEntries.java
