+++ mcedia-core/versions/1.21.8/src/client/resources/mcedia_core.client.mixins.json
